Reservation 1. Domain 수정 2. Test에서 DTO사용으로 변경 3. @Builder 사용방식으로 변경

+++ src/main/java/kr/ac/kopo/domain/PlaceVO.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PlaceVO {
-    private int id;
+    private Long id;
     private String name;
     private String location;
-    private int schoolId;
+    private Long schoolId;
 }
